【修改内容】1.增加履职计划表管理的修改和取消功能 2.完善导入月度履职计划按钮逻辑 【修改日期】2022.10.19 【修改人】伍坚山

+++ src/main/java/com/fssy/shareholder/management/pojo/system/performance/employee/EntryCasPlanDetail.java
@@ -157,13 +157,13 @@ public class EntryCasPlanDetail extends BaseModel {
      * 申报年份
      */
     @TableField("year")
-    private Long year;
+    private Integer year;
 
     /**
      * 申报月份
      */
     @TableField("month")
-    private Long month;
+    private Integer month;
 
     /**
      * 对应工作事件的计划内容
